Add files via upload
Problem Statement
AtliQ Hardware, with multiple branches across India, provides computer hardware and peripheral manufacturers to its clients. The HR department of a company is facing challenges in monitoring employee attendance and identifying patterns of absenteeism. They need to have a comprehensive view of the attendance records of all employees, including the number of leaves taken and reasons for absence.

Solution
As a data analyst, to solve the problem of analyzing employee attendance and presence for HR, developing BI dashboard can help in identifying trends, patterns, and issues that could impact the company's productivity and profitability.

Data
We are provided with the Excel datasheet which contains the employees data for 3 months between April 2022-June 2022. Data set contains employees attendance data along with attendance keys Eg: SL --> Sick Leave.

Task
•	Find Total Employees Attendance %, Sick Leave % and Work from home %
•	Find the trends for it over time
•	Analyzing the same trends in week days
Insights
•	There is a slight decrease in employees attendance over time
•	There is an increase in the trend where people opted for Work from home
•	There is an slight increase in the Sick leave trend over time
•	Most people attended office on Tuesdays
•	Most people opted for Work from home on Fridays
•	Most people took sick leaves on Mondays